A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.

A flexible version of ACCEPT 2.0 model, where the missing covairates will be imputed by using MICE approach.

accept(data, version = "flexccept", prediction_interval = FALSE, ...)

Arguments

data

new patient data with missing values to be imputed before prediction with the same format as accept samplePatients.

version

indicates which version of ACCEPT needs to be called.

prediction_interval

default is FALSE If set to true, returns prediction intervals of the predictions.

...

for other versions of accept.

Value

patientData with prediction.

Examples

results <- accept(data = samplePatients)