Creates bar graph comparing no treatment with azithromycin treatment

Creates bar graph comparing no treatment with azithromycin treatment

plotExacerbations(
  patientResults,
  type = "rate",
  interval = "PI",
  colors = c("#007bff", "rgb(204,204,204)")
)

Arguments

patientResults

patient results vector, produced by accept.

type

string: either "probability" or "rate"

interval

string: either "CI" or "PI" PI = Predicted Interval CI = Confidence Interval

colors

vector: a vector of colors to be used in the graph must be length 2 can use hexadecimal, rgb, or R color codes

Value

a bar graph

Examples

results <- accept1(samplePatients[1,])
plotExacerbations(results)