Predicts probability of observing n exacerbations in the next year

Predicts probability of observing n exacerbations in the next year

predictCountProb(patientResults, n = 10, shortened = TRUE)

Arguments

patientResults
patient results vector, produced by accept.

n
how many exacerbations

shortened
boolean: Shortened results groups into 0, 1, 2, and 3 or more exacerbations

Value

a matrix of probabilities with the number of exacerbations as rows and number of severe exacerbations as columns

Examples

results <- accept(samplePatients[1,])
predictCountProb (results)

#>                         none severe  1 severe  2 severe 3 or more severe
#> no exacerbations         0.18879524 0.0000000 0.0000000       0.00000000
#> 1 exacerbation           0.15441823 0.1603208 0.0000000       0.00000000
#> 2 exacerbations          0.06315040 0.1311286 0.0680705       0.00000000
#> 3 or more exacerbations  0.02140234 0.0718003 0.0861611       0.04365578