Predicts probability of observing n exacerbations in the next year

Predicts probability of observing n exacerbations in the next year

predictCountProb(patientResults, n = 10, shortened = TRUE)

Arguments

patientResults

patient results vector, produced by accept.

n

how many exacerbations

shortened

boolean: Shortened results groups into 0, 1, 2, and 3 or more exacerbations

Value

a matrix of probabilities with the number of exacerbations as rows and number of severe exacerbations as columns

Examples

results <- accept2(samplePatients[1,])
predictCountProb (results)
#>                         none severe   1 severe   2 severe 3 or more severe
#> no exacerbations         0.16721119 0.00000000 0.00000000       0.00000000
#> 1 exacerbation           0.11883372 0.18022310 0.00000000       0.00000000
#> 2 exacerbations          0.04222640 0.12808103 0.09712378       0.00000000
#> 3 or more exacerbations  0.01206628 0.05851757 0.10055944       0.07102149